requirements-production.txt
# Core FastAPI - Production minimal set
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & Supabase
supabase==2.0.0

# Data validation & serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & Embeddings
openai==1.3.0
anthropic==0.7.0

# Security
pyjwt[crypto]==2.8.0

# Utilities
python-dateutil==2.8.2
httpx==0.25.0

# Math operations
numpy==1.24.4

# File processing (minimal - will add document libraries later)
python-magic-bin==0.4.14  # Cross-platform magic binary
python-docx==1.1.0